feat: Extends OPL
